% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_int.R
\name{average_int}
\alias{average_int}
\title{average_int}
\usage{
average_int(dir.path = ".")
}
\arguments{
\item{dir.path}{the files direction path}
}
\value{
avg.data list; avg.result: average result for each fragment;
    int.result: intensity of all files
}
\description{
Averge intensity of same fragment in multiple file.
}
\author{
Zhiwei Zhou
}
