% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_msms_lib.R
\name{build_msms_lib}
\alias{build_msms_lib}
\title{build_msms_lib}
\usage{
build_msms_lib(dir.path, adduct = c("[M+H]", "[M+Na]", "[M+NH4]", "[M-H]",
  "[M+HCOO]"))
}
\arguments{
\item{dir.path}{directory path}

\item{adduct}{the adduct type. It supports "[M+H]", "[M+Na]", "[M+NH4]", "[M-H]", "[M+HCOO]"}
}
\author{
Zhiwei Zhou
}
