% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_MSMS_table.R
\name{cal_ion_mz}
\alias{cal_ion_mz}
\title{cal_ion_mz}
\usage{
cal_ion_mz(ext.mass, name = NULL, is.output = FALSE)
}
\arguments{
\item{ext.mass}{The exact mass of compound.}

\item{name}{The name of csv file name. Default: "ext_mass_table.csv"}
}
\description{
generate adducts mz
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
