% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_MSMS_table.R
\name{generate_MSMS_table}
\alias{generate_MSMS_table}
\title{generate_MSMS_table}
\usage{
generate_MSMS_table(com.name, ext.mass, mz = NULL, adduct = c("[M+H]",
  "[M+Na]", "[M+NH4]", "[M-H]", "[M+HCOO]"), ce = 20, charge = 1, rt = 0,
  delta.rt = 0.5, iso.width = c("Wide", "Medium", "Narrow"))
}
\arguments{
\item{com.name}{Tompound name.}

\item{ext.mass}{The exact mass of compound.}

\item{mz}{The m/z value of adduct. If it inputted, the ext.mass will be replaced. Default: NULL}

\item{adduct}{The adduct form to dissociation, including "[M+H]", "[M+Na]", "[M+NH4]","[M-H]", "[M+HCOO]"}

\item{ce}{The collision energy. Default: 20V}

\item{charge}{The charge form of ion. Default: 1}

\item{rt}{The rt range of targeted ion. Default: ""}

\item{delta.rt}{The delta rt range. Default: ""}

\item{iso.width}{The isolation window, including: "Wide", "Medium", "Narrow"}
}
\description{
generate targeted MS/MS table for each compound
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
