% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipLibBuild.R
\name{LipLibBuild}
\alias{LipLibBuild}
\title{Lipid library build}
\usage{
LipLibBuild(sdf.file, smile.file, is.output = TRUE)
}
\arguments{
\item{sdf.file}{The file name of sdf file}

\item{smile.file}{The file name of smile file}

\item{is.output}{The output csv file. The default is TRUE.}
}
\description{
Build lipid library based on sdf file and smile file
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
