% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_msms_purify.R
\name{msms_plot}
\alias{msms_plot}
\title{msms_plot}
\usage{
msms_plot(spectra.name = NULL, data.name = NULL, precusor.mz = NULL)
}
\arguments{
\item{spectra.name}{The csv file name of spectra. Default: NULL}

\item{data.name}{The R Dataname after purification.}

\item{precusor.mz}{The precusor mz. Default: NULL}
}
\description{
plot msms spectra of purified spectra
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
