% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_msms.R
\name{split_msms}
\alias{split_msms}
\title{split_msms}
\usage{
split_msms(file.name)
}
\arguments{
\item{file.name}{The csv file name for split.}
}
\description{
split msms fragments into independent csv file according to TJ lipid library. The template should be labeled as "HP"="M+H"; "NaP"="M+Na"; "NH4P"="M+NH4"; "HN"="M-H"; "HCOON"="M+HCOO". The number "1" represents the start column; number "2" represents the end column.
}
\author{
Zhiwei Zhou
{zhouzw@sioc.ac.cn}
}
