% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_subclass.R
\name{split_subclass}
\alias{split_subclass}
\title{split_subclass}
\usage{
split_subclass(raw.data = "result.csv",
  raw.MD = "Define-descriptors by rcdk.csv", bulk.std.name)
}
\arguments{
\item{raw.data}{The csv file name of raw data.}

\item{raw.MD}{The csv file name of Molecular Descriptor.}

\item{bulk.std.name}{The bulk structure name of each classes.}
}
\description{
split a big library into subclasses
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
