% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Info_Cal.R
\name{Info_Cal}
\alias{Info_Cal}
\title{Info_cal}
\usage{
Info_Cal(raw.data)
}
\arguments{
\item{raw.data}{The smiles structure as vector format.}
}
\value{
a matrix (n*3) including formula, exact mass, and charge
}
\description{
calculate formula, exact mass, and charge
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
